Rename CLI to asynq
